+++ delivery/src/main/java/com/faster/delivery/app/delivery/infrastructure/client/DeliveryManagerClientImpl.java
@@ -6,6 +6,7 @@ import com.faster.delivery.app.delivery.application.dto.DeliveryManagerDto;
 import com.faster.delivery.app.delivery.infrastructure.client.dto.deliverymanager.DeliveryManagerGetResponseDto;
 import java.util.UUID;
 import lombok.RequiredArgsConstructor;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Component;
 
 @Component
@@ -15,16 +16,17 @@ public class DeliveryManagerClientImpl implements DeliveryManagerClient {
   private final DeliveryManagerFeignClient deliveryManagerFeignClient;
 
   public DeliveryManagerDto getDeliveryManagerData(UUID deliveryManagerId) {
-    ApiResponse<DeliveryManagerGetResponseDto> deliveryManagerData =
+
+    ResponseEntity<ApiResponse<DeliveryManagerGetResponseDto>> deliveryManagerData =
         deliveryManagerFeignClient.getDeliveryManagerData(deliveryManagerId);
-    DeliveryManagerGetResponseDto data = deliveryManagerData.data();
+    DeliveryManagerGetResponseDto data = deliveryManagerData.getBody().data();
     return data.toDeliveryManagerDto();
   }
 
   @Override
   public DeliveryManagerDto assignCompanyDeliveryManager(UUID companyId) {
     return deliveryManagerFeignClient.assignCompanyDeliveryManager(companyId)
-        .data().toDeliveryManagerDto();
+        .getBody().data().toDeliveryManagerDto();
   }
 
   @Override
