+++ delivery/src/main/java/com/faster/delivery/app/delivery/infrastructure/client/MessageClientImpl.java
@@ -1,6 +1,6 @@
-package com.faster.delivery.app.deliverymanager.infrastructure.client;
+package com.faster.delivery.app.delivery.infrastructure.client;
 
-import com.faster.delivery.app.deliverymanager.application.MessageClient;
+import com.faster.delivery.app.delivery.application.MessageClient;
 import com.faster.delivery.app.deliverymanager.application.dto.SendMessageApplicationRequestDto;
 import com.faster.delivery.app.deliverymanager.infrastructure.client.dto.SendMessageFeignRequestDto;
 import lombok.RequiredArgsConstructor;
