+++ delivery/src/main/java/com/faster/delivery/app/delivery/infrastructure/client/MessageFeignClient.java
@@ -1,4 +1,4 @@
-package com.faster.delivery.app.deliverymanager.infrastructure.client;
+package com.faster.delivery.app.delivery.infrastructure.client;
 
 import com.common.response.ApiResponse;
 import com.faster.delivery.app.deliverymanager.infrastructure.client.dto.SendMessageFeignRequestDto;
